requirements.txt
# Core
requests>=2.25.0
rich>=10.0.0

# Optional
GitPython>=3.1.0
SpeechRecognition>=3.8.0
pyttsx3>=2.90
psutil>=5.8.0

# Development
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.812